requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiosmtpd==1.4.6
async-timeout==4.0.3
atpublic==5.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
frozenlist==1.4.1
idna==3.8
multidict==6.1.0
pyTelegramBotAPI==4.22.1
python-dotenv==1.1.1
requests==2.32.3
telebot==0.0.5
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.11.1
